PROMISE ECG data

PROMISE ECG data

ecg

Format

a data.table with 125462 observations on the following 23 variables:

MRN

character Medical record number

ECGDate

POSIXct ECG exam date, YYYY-MM-DD HH:MM:SS

PRInterval

numeric Interval between start of P wave and start QRS, ms

QRSDur

numeric Duration of the QRS-complex, ms

QTCorrected

numeric Corrected QT-interval according to Bazzet's formula, ms

AXIS

numeric R wave axis, degree

original_ecg_statement

character Original diagnosis statement in raw ECG XML

filepath

character Name of the raw ECG XML file

Normal_conduction

logical

Sinus_rhytm

numeric

SVT

numeric Supraventricular tachycardia (see details)

Junctional_rhythm

numeric Rythm originating from AV node or His Bundle

LBBB

numeric Left Bundle Branch Block, conventional criteria

RBBB

numeric Right Bundle Branch Block

RBBB_LAFB

logical Bifascicular Block, RBBB + LAFB

RBBB_LPFB

logical Bifascicular Block, RBBB + LPFB

IVCD

numeric Non-specific Intraventricular Conduction Delay

LAFB

numeric Left Anterior Fascicular Block

LPFB

numeric Left Posterior Fascicular Block

AFIB

numeric Atrial Fibrillation

PACE

numeric Paced rhytm, eg. pacemaker

WPW

numeric Wolf-Parkinson White, prexcitation

Exclude

numeric Diagnosis statement indicative for exclusion, (see details)

Source

The data comes from Duke University Hospital, Durham, NC, USA

Details

Date of data extraction 2017-03-31

Normal_conduction

See end vignette("promise", package = "promise") for logic

SVT

= 1 if ecg diagnosis statement during parsing contained:

Atrial reentry tachycardia

SVT

Flutter

Supra ventricular tachycardia

Atrial tachycardia

Exclude

= 1 if ecg diagnosis statement during parsing contained:

strip

lvad

idioventricular rhythm

ventricular tach

isorhyth

ventricular escape

asys

ventricular rhyt

atrial lead

stemi

See also