Merged Echo and ECG data within 30days

Merged Echo and ECG data within 30days

promiseData

Format

A data frame with 93118 observations on the following 123 variables:

MRN

character Medical record number

Echo.date

POSIXct Echo exam date, YYYY-MM-DD HH:MM:SS

EF

character Ejection Fraction incl < or > signs

EF.x

numeric Ejection Fraction, percentage units

AS.severity

a factor with levels AUTOGRAFT BIOPROSTHETIC AoV HOMOGRAFT INDETERMINATE FOR AS MECH PROSTHETIC AoV MILD AS MODERATE AS No AS OTHER PROSTHETIC AoV PROSTHETIC AoV SEVERE AS TRIVIAL AS

MS.severity

a factor with levels BIOPROSTHETIC MV INDETERMINATE FOR MS MECH PROSTHETIC MV MILD MS MODERATE MS No MS PROSTHETIC MV RING SEVERE MS TRIVIAL MS VALVE CLIP

AR.severity

a factor with levels INDETERMINATE FOR AR MILD AR MODERATE AR No AI No AR SEVERE AR TRIVIAL AR

MR.severity

a factor with levels INDETERMINATE FOR MR MILD MR MODERATE MR No MR SEVERE MR TRIVIAL MR

PAP.mm.hg

numeric Estimated Right Ventricular Systolic Pressure from Echo, mmHg

echo_indication

character Indication for Echo exam grouped in reasonable headings

MRN

character Medical record number

ECGDate

POSIXct ECG exam date, YYYY-MM-DD HH:MM:SS

VentricularRate

numeric Ventricular rate

AtrialRate

numeric Atrial rate

PRInterval

numeric Interval between start of P wave and start QRS, ms

QRSDur

numeric Duration of the QRS-complex, ms

QTInterval

numeric Interval between start of QRS and end of T wave, ms

QTCorrected

numeric Corrected QT-interval according to Bazzet's formula, ms

PAxis

numeric P wave axis, degree

AXIS

numeric R wave axis, degree

TAxis

numeric T wave axis, degree

QRSCount

numeric N QRS complexes measured

QOnset

numeric Q wave onset

QOffset

numeric Q wave offset

POnset

numeric P wave onset

POffset

numeric P wave offset

TOffset

numeric T wave offset

QTcFrederica

numeric Corrected QT-interval according to Frederica formula QTc=QR/RR(1/3)

original_ecg_statement

character Original diagnosis statement in raw ECG XML

filepath

character Name of the raw ECG XML file

Normal_conduction

logical

Sinus_rhytm

numeric

SVT

numeric Supraventricular tachycardia (see details)

Junctional_rhythm

numeric Rythm originating from AV node or His Bundle

LBBB

numeric Left Bundle Branch Block, conventional criteria

RBBB

numeric Right Bundle Branch Block

RBBB_LAFB

logical Bifascicular Block, RBBB + LAFB

RBBB_LPFB

logical Bifascicular Block, RBBB + LPFB

IVCD

numeric Non-specific Intraventricular Conduction Delay

LAFB

numeric Left Anterior Fascicular Block

LPFB

numeric Left Posterior Fascicular Block

AFIB

numeric Atrial Fibrillation

PACE

numeric Paced rhytm, eg. pacemaker

WPW

numeric Wolf-Parkinson White, prexcitation

Exclude

numeric Diagnosis statement indicative for exclusion, (see details)

Gender

a factor with levels FEMALE MALE

Race

a factor with levels Mix American.native Native.hawaiian.or.other.pacific.islander Caucasian African.american NA Asian Other

DOB

POSIXct Date Of Birth, YYYY-MM-DD HH:MM:SS

Age.at.ECG

numeric Age at time of ECG exam, years

Death.index

a factor with levels ALIVE DECEASED

DOD

POSIXct Date Of Death, YYYY-MM-DD HH:MM:SS

Alive.date

POSIXct Date of latest record of ALIVE status, YYYY-MM-DD HH:MM:SS

AF.Code

numeric ICD-code

AF.Date

POSIXct Date of diagnosis, YYYY-MM-DD HH:MM:SS

AF

numeric Atrial Fibrillation/Flutter

CAD.Code

numeric ICD-code

CAD.Date

POSIXct Date of diagnosis, YYYY-MM-DD HH:MM:SS

CAD

numeric Coronary Atery Disease

CHD.Code

numeric ICD-code

CHD.Date

POSIXct Date of diagnosis, YYYY-MM-DD HH:MM:SS

CHD

numeric Congenital Heart Disease

COPD.Code

numeric ICD-code

COPD.Date

POSIXct Date of diagnosis, YYYY-MM-DD HH:MM:SS

COPD

numeric Chronic Obstructive Pulmonary Disorder

DM.Code

numeric ICD-code

DM.Date

POSIXct Date of diagnosis, YYYY-MM-DD HH:MM:SS

DM

numeric Diabetes Mellitus

HF.Code

numeric ICD-code

HF.Date

POSIXct Date of diagnosis, YYYY-MM-DD HH:MM:SS

HF

numeric Heart Failure

HTN.Code

numeric ICD-code

HTN.Date

POSIXct Date of diagnosis, YYYY-MM-DD HH:MM:SS

HTN

numeric Hypertension

OSA.Code

numeric ICD-code

OSA.Date

POSIXct Date of diagnosis, YYYY-MM-DD HH:MM:SS

OSA

numeric Obstructive Sleep Apnea

ICD.CPT.Code

character ICD-code for device

CRT.Date

POSIXct Date of diagnosis, YYYY-MM-DD HH:MM:SS

CRT

numeric Cardiac Resynchronization Therapy

LVAD.Code

character ICD-code

LVAD.Date

POSIXct Date of diagnosis, YYYY-MM-DD HH:MM:SS

LVAD

numeric Left Ventricular Assist Device

HTx.Code

character ICD-code

HTx.Date

POSIXct Date of diagnosis, YYYY-MM-DD HH:MM:SS

HTx

numeric Heart Transplantation

ClassIII.Rx.Name

character ClassIII Antiarrythmica, Drug name +/- dosage

ClassIII.Rx.Start

POSIXct Start of prescription

ClassIII.Rx.End

POSIXct End of prescription

ClassIc.Rx.Name

character ClassIc Antiarrythmica, Drug name +/- dosage

ClassIc.Rx.Start

POSIXct Start of prescription

ClassIc.Rx.End

POSIXct End of prescription

ACEI.Rx.Name

character Angiotensin converting Enzyme Inhibitor, Drug name +/- dosage

ACEI.Rx.Start

POSIXct Start of prescription

ACEI.Rx.End

POSIXct End of prescription

Amiodarone.Rx.Name

character Amiodarone Antiarrythmica, Drug name +/- dosage

Amiodarone.Rx.Start

POSIXct Start of prescription

Amiodarone.Rx.End

POSIXct End of prescription

BB.Rx.Name

character Beta-Blocker, Drug name +/- dosage

BB.Rx.Start

POSIXct Start of prescription

BB.Rx.End

POSIXct End of prescription

ClassIII.Rx.OMR.Name

character Outpatient medication, see previous

ClassIII.Rx.OMR.Date

POSIXct Date of prescription

ClassIc.Rx.OMR.Name

character Outpatient medication, see previous

ClassIc.Rx.OMR.Date

POSIXct Date of prescription

ACEI.Rx.OMR.Name

character Outpatient medication, see previous

ACEI.Rx.OMR.Date

POSIXct Date of prescription

Amiodarone.Rx.OMR.Name

character Outpatient medication, see previous

Amiodarone.Rx.OMR.Date

POSIXct Date of prescription

BB.Rx.OMR.Name

character Outpatient medication, see previous

BB.Rx.OMR.Date

POSIXct Date of prescription

ClassIII.Rx

numeric ClassIII Antiarrythmica indicator

ClassIc.Rx

numeric ClassIc Antiarrythmica indicator

ACEI.Rx

numeric Angiotensin Converting Enzyme Inhibitor indicator

Amiodarone.Rx

numeric Amiodarone Antiarrythmica indicator

BB.Rx

numeric Beta-Blocker indicator

Na.lab.Date

POSIXct Date of Lab test, YYYY-MM-DD HH:MM:SS

Na.lab.Value

numeric Sodium value, mmol/L

K.lab.Date

POSIXct Date of Lab test, YYYY-MM-DD HH:MM:SS

K.lab.Value

numeric Potassium value, mmol/L

BNP.lab.Date

POSIXct Date of Lab test, YYYY-MM-DD HH:MM:SS

BNP.lab.Value

numeric B-type Natriuretic Peptide value, pg/mL

ProBNP.lab.Date

POSIXct Date of Lab test, YYYY-MM-DD HH:MM:SS

ProBNP.lab.Value

numeric N-terminal (NT)-pro hormone BNP value, pg/mL

GFR.lab.Date

POSIXct Date of Lab test, YYYY-MM-DD HH:MM:SS

GFR.lab.Value

numeric Glomerular Filtration Rate, ml/min/1.73m2

Source

The data comes from Duke University Hospital, Durham, NC, USA.

Details

Date of data extraction 2017-03-31

Normal_conduction

See end vignette("promise", package = "promise") for logic

SVT

= 1 if ecg diagnosis statement during parsing contained:

Atrial reentry tachycardia

SVT

Flutter

Supra ventricular tachycardia

Atrial tachycardia

Exclude

= 1 if ecg diagnosis statement during parsing contained:

strip

lvad

idioventricular rhythm

ventricular tach

isorhyth

ventricular escape

asys

ventricular rhyt

atrial lead

stemi

See also